Reset the cursor
Once tab key is pressed, place the cursor 4 spaces away from
the location where the tab key was pressed.

This handles the case of when a tab key is pressed in the middle of
an existing note and the cursor was placed at the end of the document.

fix https://github.com/standardnotes/web/issues/50