fix(app/assets): Make some buttons focusable
Divs are not focusable by default, and need a considerable amount of JavaScript to make they behave similar to the native button element. This switches buttons implemented as divs to the native button element (styles are updated accordingly to maintain the same visual style). Link tags without href are also not focusable - for most of these cases the button element should be used, but that would involve a more significant refactoring, so I have opted to adding a dummy href.

While enabling keyboard navigation, this also addresses the fact that screen reader users had no idea that these buttons existed.